EVAKA-4028: combine assistance reports
